Added proper shooting point, changed Point to floats in Rust

+++ game-5-sim/src/physics.rs
@@ -4,7 +4,7 @@ pub type Velocity = f32;
 
 pub fn move_point (left_top: &Point, velocity: &Velocity, direction: &Radians) -> Point {
   return Point {
-    x: left_top.x + ((velocity * direction.cos()) as i32),
-    y: left_top.y + ((velocity * direction.sin()) as i32)
+    x: left_top.x + velocity * direction.cos(),
+    y: left_top.y + velocity * direction.sin()
   }
 }
